onnx/requirements.txt
flatbuffers==2.0
numpy==1.21.4
onnxruntime==1.9.0
Pillow==8.4.0
protobuf==3.19.1
